requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.8
django-heroku==0.3.1
django-taggit==1.1.0
django-widget-tweaks==1.4.5
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
python-decouple==3.3
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.4
